refactor(render): move the family warm-up out of PackChain
Beside the frame, PackChain held the whole of the pack-load warm-up: the
pool of three daemon threads the family compiles run on, the set of
workers still in flight that shutdown waits out, the flag that stops a
worker for a chain nothing will ever draw again, and the counters the
workers keep as they walk, two of which the compile card's words are
made of. It is one mechanism, it answers to nothing else, and it sat in
the middle of a class that is otherwise about drawing a frame.

It becomes FamilyWarmup, one per chain, built with the six families, the
archive they are read from and the lock over their program maps. The
chain keeps that lock, the emptying it guards being its own, and reaches
what it still needs through familiesReady, familyPrograms, warmed,
warmedAt, walked and total. A worker reads the stopped pack through
PackChain.stopped rather than off the field.
